+++ src/bfactory/core/templates/helpers/field_type_linting.py
@@ -4,4 +4,4 @@
 {%- if field.type == 'int' or field.type == 'pint' -%}int{% endif %}
 {%- if field.type == 'fk' -%}{{field.fk}}{% endif %}
 {%- if field.type == 'datetime' -%}datetime{% endif %}
-{%- if field.type == 'float' -%}float{% endif %}
+{%- if field.type == 'decimal' -%}float{% endif %}
